+++ crates/integrations/connector-integration/src/connectors/airwallex/transformers.rs
@@ -613,6 +613,7 @@ impl<T: PaymentMethodDataTypes> TryFrom<ResponseRouterData<AirwallexPaymentsResp
                 mandate_reference: None,
                 connector_metadata,
                 network_txn_id,
+                network_txn_link_id: None,
                 connector_response_reference_id: item.response.payment_intent_id,
                 incremental_authorization_allowed: Some(false), // Airwallex doesn't support incremental auth
                 status_code: item.http_code,
@@ -653,6 +654,7 @@ impl TryFrom<ResponseRouterData<AirwallexSyncResponse, Self>>
                 mandate_reference: None,
                 connector_metadata: None,
                 network_txn_id,
+                network_txn_link_id: None,
                 connector_response_reference_id: Some(intent_id.clone()),
                 incremental_authorization_allowed: None,
                 status_code: item.http_code,
@@ -759,6 +761,7 @@ impl TryFrom<ResponseRouterData<AirwallexCaptureResponse, Self>>
                 mandate_reference: None,
                 connector_metadata,
                 network_txn_id,
+                network_txn_link_id: None,
                 connector_response_reference_id: item.response.payment_intent_id,
                 incremental_authorization_allowed: Some(false), // Airwallex doesn't support incremental auth
                 status_code: item.http_code,
@@ -1015,6 +1018,7 @@ impl TryFrom<ResponseRouterData<AirwallexVoidResponse, Self>>
                 mandate_reference: None,
                 connector_metadata,
                 network_txn_id,
+                network_txn_link_id: None,
                 connector_response_reference_id: item.response.payment_intent_id,
                 incremental_authorization_allowed: Some(false), // Airwallex doesn't support incremental auth
                 status_code: item.http_code,
@@ -1614,6 +1618,7 @@ impl<T: PaymentMethodDataTypes> TryFrom<ResponseRouterData<AirwallexSetupMandate
                 mandate_reference,
                 connector_metadata: None,
                 network_txn_id: None,
+                network_txn_link_id: None,
                 connector_response_reference_id: item.response.payment_intent_id,
                 incremental_authorization_allowed: Some(false),
                 status_code: item.http_code,
@@ -1762,6 +1767,7 @@ impl<T: PaymentMethodDataTypes> TryFrom<ResponseRouterData<AirwallexRepeatPaymen
                 mandate_reference: None,
                 connector_metadata: None,
                 network_txn_id: None,
+                network_txn_link_id: None,
                 connector_response_reference_id: item.response.payment_intent_id,
                 incremental_authorization_allowed: Some(false),
                 status_code: item.http_code,
@@ -1873,6 +1879,7 @@ impl TryFrom<ResponseRouterData<AirwallexCustomerResponse, Self>>
         let mut router_data = item.router_data;
         router_data.response = Ok(ConnectorCustomerResponse {
             connector_customer_id: item.response.id,
+            status_code: item.http_code,
         });
         router_data.resource_common_data.connector_http_status_code = Some(item.http_code);
         Ok(router_data)