+++ crates/integrations/connector-integration/src/connectors/rapyd/transformers.rs
@@ -145,6 +145,7 @@ impl<F, T> TryFrom<ResponseRouterData<RapydPaymentsResponse, Self>>
                                 mandate_reference: None,
                                 connector_metadata: None,
                                 network_txn_id: None,
+                                network_txn_link_id: None,
                                 connector_response_reference_id: data
                                     .merchant_reference_id
                                     .to_owned(),
@@ -1296,6 +1297,7 @@ impl<T: PaymentMethodDataTypes + Debug + Sync + Send + 'static + Serialize>
                                         mandate_reference,
                                         connector_metadata: None,
                                         network_txn_id: None,
+                                        network_txn_link_id: None,
                                         connector_response_reference_id: data
                                             .merchant_reference_id
                                             .clone(),
@@ -1538,6 +1540,7 @@ impl<T: PaymentMethodDataTypes + Debug + Sync + Send + 'static + Serialize>
                             mandate_reference: None,
                             connector_metadata: None,
                             network_txn_id: None,
+                            network_txn_link_id: None,
                             connector_response_reference_id: data.merchant_reference_id.to_owned(),
                             incremental_authorization_allowed: None,
                             status_code: item.http_code,