+++ crates/internal/composite-service/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod events;
+pub mod payment_methods;
 pub mod payments;
 pub mod transformers;
 pub mod utils;